+++ itests/org.jupnp.common/src/main/resources/org/jupnp/common/data/initial-data.properties
