+++ itests/org.jupnp.device.simple/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.jupnp.pom</groupId>
+    <artifactId>itests</artifactId>
+    <version>3.0.0-SNAPSHOT</version>
+  </parent>
+
+  <groupId>org.jupnp.itests</groupId>
+  <artifactId>org.jupnp.device.simple</artifactId>
+
+  <name>jUPnP Integration Tests Simple Device</name>
+  <description>A test device that exposes all the capabilities of OSGi UPnP</description>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.jupnp.itests</groupId>
+      <artifactId>org.jupnp.common</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+  </dependencies>
+
+</project>
