+++ itests/org.jupnp.device.simple/src/main/resources/org/jupnp/device/simple/icons/upnp-icon-small.png
