+++ pom.xml
@@ -36,7 +36,7 @@
   <modules>
     <module>bom</module>
     <module>bundles</module>
-    <module>tests</module>
+    <module>itests</module>
     <module>tools</module>
   </modules>
