--- tests/common/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.jupnp.pom</groupId>
-    <artifactId>tests</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.jupnp.tests</groupId>
-  <artifactId>jupnp-osgi-tests-common</artifactId>
-
-  <name>jUPnP OSGi Test Common Library</name>
-
-</project>
