--- tests/integration/itest.bndrun
@@ -1,83 +1,0 @@
--standalone: \
-	../../bom/runtime-index/target/index.xml;name="org.jupnp.bom.runtime-index",\
-	../../bom/test-index/target/index.xml;name="org.jupnp.bom.test-index",\
-	target/index.xml;name="self"
-
--resolve.effective: active
-
--tester: biz.aQute.tester.junit-platform
-
-# Run all integration tests which are named xyzTest
-Test-Cases: ${classes;CONCRETE;PUBLIC;NAMED;*Test}
-
--runfw: org.eclipse.osgi
--runee: JavaSE-11
-
-# The integration test itself does not export anything.
-Export-Package:
--exportcontents:
-
--runrequires.container: \
-	bnd.identity;id='org.osgi.service.event',\
-	bnd.identity;id='org.eclipse.equinox.event',\
-	bnd.identity;id='org.ops4j.pax.web.pax-web-jetty',\
-	bnd.identity;id='org.ops4j.pax.web.pax-web-runtime'
-
--runrequires.junit: \
-	bnd.identity;id='biz.aQute.tester.junit-platform',\
-	bnd.identity;id='junit-jupiter-engine'
-
--runproperties: \
-	org.ops4j.pax.logging.DefaultServiceLog.level=WARN
-
-Bundle-SymbolicName: ${project.groupId}.${project.artifactId}
-
--runrequires: \
-	bnd.identity;id='org.jupnp.tests.jupnp-osgi-tests-integration',\
-	bnd.identity;id='org.jupnp.osgi'
-
--runblacklist: \
-	bnd.identity;id='osgi.cmpn',\
-	bnd.identity;id='slf4j.api',\
-	bnd.identity;id='slf4j.simple'
-
-#
-# done
-#
--runbundles: \
-	biz.aQute.tester.junit-platform;version='[6.4.0,6.4.1)',\
-	javax.servlet-api;version='[3.1.0,3.1.1)',\
-	junit-jupiter-api;version='[5.10.0,5.10.1)',\
-	junit-jupiter-engine;version='[5.10.0,5.10.1)',\
-	junit-platform-commons;version='[1.10.0,1.10.1)',\
-	junit-platform-engine;version='[1.10.0,1.10.1)',\
-	junit-platform-launcher;version='[1.10.0,1.10.1)',\
-	org.apache.felix.configadmin;version='[1.9.26,1.9.27)',\
-	org.eclipse.equinox.event;version='[1.4.300,1.4.301)',\
-	org.eclipse.jetty.client;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.http;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.io;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.jaas;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.security;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.server;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.servlet;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.util;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.util.ajax;version='[9.4.52,9.4.53)',\
-	org.eclipse.jetty.xml;version='[9.4.52,9.4.53)',\
-	org.jupnp;version='[3.0.0,3.0.1)',\
-	org.jupnp.osgi;version='[3.0.0,3.0.1)',\
-	org.jupnp.osgi.upnp.test.device.simple;version='[3.0.0,3.0.1)',\
-	org.jupnp.tests.jupnp-osgi-tests-integration;version='[3.0.0,3.0.1)',\
-	org.opentest4j;version='[1.3.0,1.3.1)',\
-	org.ops4j.pax.logging.pax-logging-api;version='[2.2.4,2.2.5)',\
-	org.ops4j.pax.web.pax-web-api;version='[8.0.22,8.0.23)',\
-	org.ops4j.pax.web.pax-web-jetty;version='[8.0.22,8.0.23)',\
-	org.ops4j.pax.web.pax-web-runtime;version='[8.0.22,8.0.23)',\
-	org.ops4j.pax.web.pax-web-spi;version='[8.0.22,8.0.23)',\
-	org.ops4j.pax.web.pax-web-tomcat-common;version='[8.0.22,8.0.23)',\
-	org.osgi.service.component;version='[1.5.0,1.5.1)',\
-	org.osgi.service.upnp;version='[1.2.1,1.2.2)',\
-	org.osgi.util.function;version='[1.0.0,1.0.1)',\
-	org.osgi.util.promise;version='[1.0.0,1.0.1)',\
-	org.apache.felix.scr;version='[2.2.6,2.2.7)',\
-	org.osgi.service.event;version='[1.4.1,1.4.2)'
