--- tests/integration/pom.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.jupnp.pom</groupId>
-    <artifactId>tests</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.jupnp.tests</groupId>
-  <artifactId>jupnp-osgi-tests-integration</artifactId>
-
-  <name>jUPnP OSGi Test Integration</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.jupnp</groupId>
-      <artifactId>org.jupnp</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.jupnp</groupId>
-      <artifactId>org.jupnp.osgi</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.jupnp.tests</groupId>
-      <artifactId>org.jupnp.osgi.upnp.test.device.simple</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-jar-plugin</artifactId>
-        <configuration>
-          <archive>
-            <manifestFile>${project.build.outputDirectory}/META-INF/MANIFEST.MF</manifestFile>
-          </archive>
-          <skipIfEmpty>true</skipIfEmpty>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>biz.aQute.bnd</groupId>
-        <artifactId>bnd-maven-plugin</artifactId>
-        <configuration>
-          <bndfile>itest.bndrun</bndfile>
-        </configuration>
-        <executions>
-          <execution>
-            <goals>
-              <goal>bnd-process</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>biz.aQute.bnd</groupId>
-        <artifactId>bnd-indexer-maven-plugin</artifactId>
-        <configuration>
-          <localURLs>REQUIRED</localURLs>
-          <attach>false</attach>
-        </configuration>
-        <executions>
-          <execution>
-            <id>index</id>
-            <goals>
-              <goal>index</goal>
-            </goals>
-            <configuration>
-              <indexName>${project.artifactId}</indexName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>test-index</id>
-            <goals>
-              <goal>index</goal>
-            </goals>
-            <configuration>
-              <includeJar>true</includeJar>
-              <indexName>${project.artifactId}</indexName>
-              <outputFile>${project.build.directory}/test-index.xml</outputFile>
-              <scopes>
-                <scope>test</scope>
-              </scopes>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>biz.aQute.bnd</groupId>
-        <artifactId>bnd-testing-maven-plugin</artifactId>
-        <configuration>
-          <bndruns>
-            <bndrun>itest.bndrun</bndrun>
-          </bndruns>
-        </configuration>
-        <executions>
-          <execution>
-            <goals>
-              <goal>testing</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>biz.aQute.bnd</groupId>
-        <artifactId>bnd-resolver-maven-plugin</artifactId>
-        <version>${bnd.version}</version>
-        <configuration>
-          <failOnChanges>false</failOnChanges>
-          <bndruns></bndruns>
-        </configuration>
-        <executions>
-          <execution>
-            <goals>
-              <goal>resolve</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
