--- tests/pom.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.jupnp</groupId>
-    <artifactId>jupnp</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.jupnp.pom</groupId>
-  <artifactId>tests</artifactId>
-  <packaging>pom</packaging>
-
-  <name>jUPnP Tests</name>
-
-  <modules>
-    <module>common</module>
-    <module>devices/simple</module>
-    <module>integration</module>
-  </modules>
-
-  <properties>
-    <maven.deploy.skip>true</maven.deploy.skip>
-    <maven.javadoc.skip>true</maven.javadoc.skip>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.jupnp.bom</groupId>
-      <artifactId>org.jupnp.bom.compile</artifactId>
-      <type>pom</type>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jupnp.bom</groupId>
-      <artifactId>org.jupnp.bom.test</artifactId>
-      <type>pom</type>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jupnp.bom</groupId>
-      <artifactId>org.jupnp.bom.runtime</artifactId>
-      <type>pom</type>
-      <scope>compile</scope>
-    </dependency>
-  </dependencies>
-
-  <profiles>
-    <profile>
-      <id>with-bnd-resolver-resolve</id>
-      <activation>
-        <property>
-          <name>withResolver</name>
-        </property>
-      </activation>
-      <build>
-        <pluginManagement>
-          <plugins>
-            <plugin>
-              <groupId>biz.aQute.bnd</groupId>
-              <artifactId>bnd-resolver-maven-plugin</artifactId>
-              <version>${bnd.version}</version>
-              <executions>
-                <execution>
-                  <goals>
-                    <goal>resolve</goal>
-                  </goals>
-                  <phase>package</phase>
-                </execution>
-              </executions>
-            </plugin>
-          </plugins>
-        </pluginManagement>
-      </build>
-    </profile>
-  </profiles>
-
-</project>
